requirements.txt
telethon>=1.34.0
fyers-apiv3>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0
flask-login>=0.6.3
